Align Java entities with database schema
 Entity
Customer.java
customer_id      Integer
phone_number     String
first_name       String
last_name        String
customer_email   String
address_id       String

address_id là String - vì DB là varchar(100).

CustomerOrder.java
order_id      Integer
customer_id   Integer
status        String
order_date    LocalDateTime
total_price   BigDecimal
Payment.java
payment_id       Integer
order_id         Integer
payment_method   String
payment_status   String
payment_date     LocalDateTime
OrderDetail.java
order_detail_id  Integer
order_id         Integer
cake_id          Integer
quantity         Integer
sub_total        BigDecimal
Product.java
cake_id          Integer
option_cake_id   Integer
quantity         Integer
price            BigDecimal
description      String[]
cake_name        String
create_at        LocalDateTime
OptionCake.java
option_cake_id   Integer
category_name    String
Review.java
reviews_id        Integer
order_detail_id   Integer
created_at        LocalDateTime
customer_comment  String
rating            Integer

+++ webbanhngot/backendbanh/src/webbanhngot/entity/OptionCake.java
@@ -1,23 +1,23 @@
 package webbanhngot.entity;
 
 public class OptionCake {
-	private Long option_cake_id;
+	private Integer option_cake_id;
 	private String category_name;
 
 	public OptionCake() {
 
 	}
 
-	public OptionCake(Long option_cake_id, String category_name) {
+	public OptionCake(Integer option_cake_id, String category_name) {
 		this.option_cake_id = option_cake_id;
 		this.category_name = category_name;
 	}
 
-	public Long getOption_cake_id() {
+	public Integer getOption_cake_id() {
 		return option_cake_id;
 	}
 
-	public void setOption_cake_id(Long option_cake_id) {
+	public void setOption_cake_id(Integer option_cake_id) {
 		this.option_cake_id = option_cake_id;
 	}
 
@@ -30,7 +30,7 @@ public class OptionCake {
 	}
 
 	@Override
-public String toString() {
-    return "OptionCake [option_cake_id=" + option_cake_id + ", category_name=" + category_name + "]";
-}
+	public String toString() {
+		return "OptionCake [option_cake_id=" + option_cake_id + ", category_name=" + category_name + "]";
+	}
 }
